requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.18
chromadb==1.0.7
pandas==2.2.3
tiktoken==0.9.0
python-dotenv==1.1.0
